console: Streamline device store files
