+++ pkg/webui/console/store/middleware/logics/devices.js
@@ -15,7 +15,36 @@
 import api from '../../../api'
 import * as devices from '../../actions/devices'
 import * as deviceTemplateFormats from '../../actions/device-template-formats'
+import createEventsConnectLogics from './events'
 import createRequestLogic from './lib'
+
+const getDeviceLogic = createRequestLogic({
+  type: devices.GET_DEV,
+  async process({ action }, dispatch) {
+    const {
+      payload: { appId, deviceId },
+      meta: { selector, options },
+    } = action
+    const dev = await api.device.get(appId, deviceId, selector, options)
+    dispatch(devices.startDeviceEventsStream(dev.ids))
+    return dev
+  },
+})
+
+const updateDeviceLogic = createRequestLogic(
+  {
+    type: devices.UPDATE_DEV,
+    async process({ action }) {
+      const {
+        payload: { appId, deviceId, patch },
+      } = action
+      const result = await api.device.update(appId, deviceId, patch)
+
+      return { ...patch, ...result }
+    },
+  },
+  devices.updateDeviceSuccess,
+)
 
 const getDevicesListLogic = createRequestLogic({
   type: devices.GET_DEVICES_LIST,
@@ -39,4 +68,10 @@ const getDeviceTemplateFormatsLogic = createRequestLogic({
   },
 })
 
-export default [getDevicesListLogic, getDeviceTemplateFormatsLogic]
+export default [
+  getDevicesListLogic,
+  getDeviceTemplateFormatsLogic,
+  getDeviceLogic,
+  updateDeviceLogic,
+  ...createEventsConnectLogics(devices.SHARED_NAME, 'devices', api.device.eventsSubscribe),
+]
