+++ pkg/webui/console/store/middleware/logics/index.js
@@ -17,7 +17,6 @@ import users from './users'
 import init from './init'
 import applications from './applications'
 import devices from './devices'
-import device from './device'
 import gateways from './gateways'
 import configuration from './configuration'
 import organizations from './organizations'
@@ -29,7 +28,6 @@ export default [
   ...init,
   ...applications,
   ...devices,
-  ...device,
   ...gateways,
   ...configuration,
   ...organizations,
