--- pkg/webui/console/store/reducers/device.js
@@ -1,37 +1,0 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import { merge } from 'lodash'
-import { GET_DEV_SUCCESS, UPDATE_DEV_SUCCESS } from '../actions/device'
-
-const defaultState = {
-  device: undefined,
-}
-
-const device = function(state = defaultState, { type, payload }) {
-  switch (type) {
-    case UPDATE_DEV_SUCCESS:
-      const mergedDevice = merge({}, state.device, payload)
-      return { device: mergedDevice }
-    case GET_DEV_SUCCESS:
-      return {
-        ...state,
-        device: payload,
-      }
-    default:
-      return state
-  }
-}
-
-export default device
