+++ pkg/webui/console/views/device-data/index.js
@@ -23,7 +23,8 @@ import Breadcrumb from '../../../components/breadcrumbs/breadcrumb'
 import { withBreadcrumb } from '../../../components/breadcrumbs/context'
 import DeviceEvents from '../../containers/device-events'
 
-import { getDeviceId } from '../../../lib/selectors/id'
+import { selectSelectedDevice, selectSelectedDeviceId } from '../../store/selectors/devices'
+
 import PropTypes from '../../../lib/prop-types'
 
 @connect(function(state, props) {
